YAN-27 Рефакторинг - вынес повторяющуюся логику в хук
