+++ packages/client/src/components/Levels/Engine/hooks/index.ts
@@ -1,3 +1,4 @@
 export { default as useMousePosition } from './useMousePosition'
 export { default as usePlacementTiles } from './usePlacementTiles'
+export { default as useStartGame } from './useStartGame'
 export { default as useActiveTile } from './useActiveTile'
